Revert "Rename AudioDeviceManager to AudioManagerWrapper"
This reverts commit 72c6194d632603af58569574386454179a44245b.
