+++ audioswitch/src/main/java/com/twilio/audioswitch/bluetooth/EnableBluetoothScoJob.kt
@@ -2,20 +2,20 @@ package com.twilio.audioswitch.bluetooth
 
 import android.os.Handler
 import android.os.Looper
-import com.twilio.audioswitch.android.AudioManagerWrapper
 import com.twilio.audioswitch.android.LogWrapper
 import com.twilio.audioswitch.android.SystemClockWrapper
 import com.twilio.audioswitch.bluetooth.BluetoothDeviceConnectionListener.ConnectionError.SCO_CONNECTION_ERROR
+import com.twilio.audioswitch.selection.AudioDeviceManager
 
 internal class EnableBluetoothScoJob(
     logger: LogWrapper,
-    private val audioManagerWrapper: AudioManagerWrapper,
+    private val audioDeviceManager: AudioDeviceManager,
     bluetoothScoHandler: Handler = Handler(Looper.getMainLooper()),
     systemClockWrapper: SystemClockWrapper = SystemClockWrapper()
 ) : BluetoothScoJob(logger, bluetoothScoHandler, systemClockWrapper) {
 
     override val scoAction = {
-        audioManagerWrapper.enableBluetoothSco(true)
+        audioDeviceManager.enableBluetoothSco(true)
     }
 
     override val timeoutError = SCO_CONNECTION_ERROR
